code/requirements.txt
openai
python-dotenv
datasets
pandas
numpy
matplotlib
seaborn
scikit-learn
